--- a/backend/src/graphql/resolver/BalanceResolver.ts
+++ b/backend/src/graphql/resolver/BalanceResolver.ts
@@ -29,11 +29,18 @@ export class BalanceResolver {
 
     const gdtResolver = new GdtResolver()
     const balanceGDT = await gdtResolver.gdtBalance(context)
+
+    logger.info(`time for load gdt balance: ${new Date().getTime() - now.getTime()} ms`)
+    let profilingTime = new Date()
+
     logger.debug(`balanceGDT=${balanceGDT}`)
 
     const lastTransaction = context.lastTransaction
       ? context.lastTransaction
       : await getLastTransaction(user.id)
+
+    logger.info(`time for load lastTransaction from db (if not already done): ${new Date().getTime() - profilingTime.getTime()} ms`)
+    profilingTime = new Date()
 
     logger.debug(`lastTransaction=${lastTransaction}`)
 
@@ -52,7 +59,11 @@ export class BalanceResolver {
       context.transactionCount || context.transactionCount === 0
         ? context.transactionCount
         : await dbTransaction.count({ where: { userId: user.id } })
+    
     logger.debug(`transactionCount=${count}`)
+
+    logger.info(`time for count transaction in db: ${new Date().getTime() - profilingTime.getTime()} ms`)
+    profilingTime = new Date()    
 
     const linkCount = await dbTransactionLink.count({
       where: {
@@ -62,6 +73,9 @@ export class BalanceResolver {
       },
     })
     logger.debug(`linkCount=${linkCount}`)
+
+    logger.info(`time for count transaction links in db: ${new Date().getTime() - profilingTime.getTime()} ms`)
+    profilingTime = new Date()
 
     // The decay is always calculated on the last booked transaction
     const calculatedDecay = calculateDecay(
@@ -80,6 +94,9 @@ export class BalanceResolver {
     const { sumHoldAvailableAmount } = context.sumHoldAvailableAmount
       ? { sumHoldAvailableAmount: context.sumHoldAvailableAmount }
       : await transactionLinkSummary(user.id, now)
+
+    logger.info(`time for load transactionLinkSummary from db (if not already done): ${new Date().getTime() - profilingTime.getTime()} ms`)
+    profilingTime = new Date()
 
     logger.debug(`context.sumHoldAvailableAmount=${context.sumHoldAvailableAmount}`)
     logger.debug(`sumHoldAvailableAmount=${sumHoldAvailableAmount}`)