--- a/backend/src/graphql/resolver/TransactionResolver.ts
+++ b/backend/src/graphql/resolver/TransactionResolver.ts
@@ -38,6 +38,7 @@ import { calculateBalance } from '@/util/validate'
 import { virtualLinkTransaction, virtualDecayTransaction } from '@/util/virtualTransactions'
 
 import { BalanceResolver } from './BalanceResolver'
+import { GdtResolver } from './GdtResolver'
 import { getCommunityByIdentifier, getCommunityName, isHomeCommunity } from './util/communities'
 import { findUserByIdentifier } from './util/findUserByIdentifier'
 import { getLastTransaction } from './util/getLastTransaction'
@@ -49,7 +50,6 @@ import {
 import { sendTransactionsToDltConnector } from './util/sendTransactionsToDltConnector'
 import { storeForeignUser } from './util/storeForeignUser'
 import { transactionLinkSummary } from './util/transactionLinkSummary'
-import { GdtResolver } from './GdtResolver'
 
 export const executeTransaction = async (
   amount: Decimal,
@@ -448,7 +448,7 @@ export class TransactionResolver {
     profilingTime = new Date()
     const balanceGDT = await balanceGDTPromise
     context.balanceGDT = balanceGDT
-    
+
     // Construct Result
     return new TransactionList(await balanceResolver.balance(context), transactions)
   }