--- a/backend/src/graphql/resolver/TransactionResolver.ts
+++ b/backend/src/graphql/resolver/TransactionResolver.ts
@@ -237,8 +237,6 @@ export class TransactionResolver {
     const lastTransaction = await getLastTransaction(user.id)
     logger.debug(`lastTransaction=${lastTransaction}`)
 
-    logger.info(`time for get last transaction: ${new Date().getTime() - now.getTime()} ms`)
-
     const balanceResolver = new BalanceResolver()
     context.lastTransaction = lastTransaction
 
@@ -257,8 +255,6 @@ export class TransactionResolver {
       order,
     )
     context.transactionCount = userTransactionsCount
-    let profilingTime = new Date()
-    logger.info(`time for getTransactionList: ${profilingTime.getTime() - now.getTime()} ms`)
 
     // find involved users; I am involved
     const involvedUserIds: number[] = [user.id]
@@ -311,14 +307,7 @@ export class TransactionResolver {
       }
     }
     logger.debug(`involvedUserIds=`, involvedUserIds)
-    logger.debug(`involvedRemoteUsers=`, involvedRemoteUsers)
-
-    logger.info(
-      `time for collect involved user and load remote user: ${
-        new Date().getTime() - profilingTime.getTime()
-      } ms`,
-    )
-    profilingTime = new Date()
+    logger.debug(`involvedRemoteUsers=`, involvedRemoteUsers)    
 
     // We need to show the name for deleted users for old transactions
     const involvedDbUsers = await dbUser.find({
@@ -328,23 +317,12 @@ export class TransactionResolver {
     })
     const involvedUsers = involvedDbUsers.map((u) => new User(u))
     logger.debug(`involvedUsers=`, involvedUsers)
-    logger.info(
-      `time for load involved user from db: ${new Date().getTime() - profilingTime.getTime()} ms`,
-    )
-    profilingTime = new Date()
-
+    
     const self = new User(user)
     const transactions: Transaction[] = []
 
     const { sumHoldAvailableAmount, sumAmount, lastDate, firstDate, transactionLinkcount } =
       await transactionLinkSummary(user.id, now)
-
-    logger.debug(
-      `time for load transactionLinkSummary db: ${
-        new Date().getTime() - profilingTime.getTime()
-      } ms`,
-    )
-    profilingTime = new Date()
 
     context.linkCount = transactionLinkcount
     logger.debug(`transactionLinkcount=${transactionLinkcount}`)
@@ -440,12 +418,6 @@ export class TransactionResolver {
         ).toDecimalPlaces(2, Decimal.ROUND_HALF_UP)
       }
     })
-    logger.info(
-      `time for process transaction list and fill in decay db: ${
-        new Date().getTime() - profilingTime.getTime()
-      } ms`,
-    )
-    profilingTime = new Date()
     const balanceGDT = await balanceGDTPromise
     context.balanceGDT = balanceGDT