--- a/backend/src/logging/BalanceLogging.view.ts
+++ b/backend/src/logging/BalanceLogging.view.ts
@@ -1,0 +1,18 @@
+import { Balance } from '@/graphql/model/Balance'
+import { AbstractLoggingView } from '@logging/AbstractLogging.view'
+
+export class BalanceLoggingView extends AbstractLoggingView {
+  public constructor(private self: Balance) {
+    super()
+  }
+
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  public toJSON(): any {
+    return {
+      balance: this.decimalToString(this.self.balance),
+      balanceGDT: this.self.balanceGDT,
+      count: this.self.count,
+      linkCount: this.self.linkCount,
+    }
+  }
+}