--- a/backend/src/logging/BalanceLogging.view.ts
+++ b/backend/src/logging/BalanceLogging.view.ts
@@ -1,5 +1,6 @@
-import { Balance } from '@/graphql/model/Balance'
 import { AbstractLoggingView } from '@logging/AbstractLogging.view'
+
+import { Balance } from '@/graphql/model/Balance'
 
 export class BalanceLoggingView extends AbstractLoggingView {
   public constructor(private self: Balance) {