--- a/database/migrations/0085-add_index_transactions_user_id.ts
+++ b/database/migrations/0085-add_index_transactions_user_id.ts
@@ -1,0 +1,10 @@
+/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
+/* eslint-disable @typescript-eslint/no-explicit-any */
+
+export async function upgrade(queryFn: (query: string, values?: any[]) => Promise<Array<any>>) {
+  await queryFn('CREATE INDEX user_id ON transactions (user_id);')
+}
+
+export async function downgrade(queryFn: (query: string, values?: any[]) => Promise<Array<any>>) {
+  await queryFn('DROP INDEX user_id ON transactions;')
+}