<!-- You can find the latest issue templates here https://github.com/ulfgebhardt/issue-templates -->

## 🍰 Pullrequest
<!-- Describe the Pullrequest. Use Screenshots if possible. -->
- make TransactionResolver.transactionList faster, improvement on stage2.gradido.net from ~350 ms to ~100 ms with production data set copy
- calling gdt server for gdt sum parallel to rest of transactionList logic
  - calling seems to be needing at least 100 ms, independent of gdt server speed (return balance in 15 ms)
- adding INDEX for transactions.user_id improve speed for getLastTransaction and getTransactionList dramatic
  - getLastTransaction from ~127 ms to 4 ms
  - getTransactionList ~300 ms to 14 ms
- reduce log output from decimal in BalanceResolver.balance with added DecayLoggingView and BalanceLoggingView